+++ crates/utiles/src/lager/reloadable.rs
@@ -1,6 +1,7 @@
-use once_cell::sync::Lazy;
 use std::io;
+use std::sync::LazyLock;
 use std::sync::Mutex;
+
 use tracing::debug;
 use tracing_subscriber::fmt;
 use tracing_subscriber::layer::{Layered, SubscriberExt};
@@ -22,12 +23,12 @@ type LagerFormatLayer = Handle<
     Layered<Box<dyn Layer<Registry> + Send + Sync>, Registry, Registry>,
 >;
 // Handle<Box<dyn Layer<Registry> + Send + Sync>, Registry>;
-pub static GLOBAL_FILTER_RELOAD_HANDLE: Lazy<Mutex<Option<LagerLayer>>> =
-    Lazy::new(|| Mutex::new(None));
-pub static GLOBAL_FORMAT_RELOAD_HANDLE: Lazy<Mutex<Option<LagerFormatLayer>>> =
-    Lazy::new(|| Mutex::new(None));
-pub static GLOBAL_LAGER_CONFIG: Lazy<Mutex<LagerConfig>> =
-    Lazy::new(|| Mutex::new(LagerConfig::default()));
+pub static GLOBAL_FILTER_RELOAD_HANDLE: LazyLock<Mutex<Option<LagerLayer>>> =
+    LazyLock::new(|| Mutex::new(None));
+pub static GLOBAL_FORMAT_RELOAD_HANDLE: LazyLock<Mutex<Option<LagerFormatLayer>>> =
+    LazyLock::new(|| Mutex::new(None));
+pub static GLOBAL_LAGER_CONFIG: LazyLock<Mutex<LagerConfig>> =
+    LazyLock::new(|| Mutex::new(LagerConfig::default()));
 
 // #[derive(Debug, Default, Copy, Clone)]
 // pub struct LagerConfig {